Moves pinman package to pins so packages agree in name with thier subcommand name
