+++ src/apps/chifra/cmd/pins/handle_list.go
@@ -10,7 +10,7 @@
  * General Public License for more details. You should have received a copy of the GNU General
  * Public License along with this program. If not, see http://www.gnu.org/licenses/.
  *-------------------------------------------------------------------------------------------*/
-package pinman
+package pins
 
 import (
 	"fmt"
